+++ CHANGELOG.md
@@ -27,6 +27,9 @@
 ### opscode-omnibus
 * Add ability to configure SQL query timeout for Erchef, bifrost and mover.
 * Provide reasonable default for LDAP and LDAPS ports.
+* Deprecate ldap "encryption" setting and replace with
+  `ssl_enabled`/`tls_enabled`. Add further validation and sanity checks around
+  ldap settings, as well as deprecation warnings.
 * Add ability to configure timeout for connect() when connecting to backends.
 
 ### oc\_erchef 0.29.4